student-management-project/backend/requirements.txt
Flask==2.3.2
pymongo==4.4.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
